+++ RazorSvelte/App/About.svelte
@@ -1,0 +1,23 @@
+<script lang="ts">
+    import Paper, { Title, Content } from '@smui/paper';
+    import Layout from "./Shared/Layout.svelte";
+    
+</script>
+
+<Layout>
+    <Paper color="primary" variant="outlined" style="margin-top: 25px">
+        <Title>
+            <code>Norm.net</code>
+        </Title>
+        <Content>
+            <p>
+                High performance micro-ORM database mapper and modernized Dapper replacement for .NET Standard 2.1 and higher
+            </p>
+            <p>
+                Check it out on <a href="https://github.com/vb-consulting/Norm.net">GitHub</a>
+            </p>
+        </Content>
+    </Paper>
+</Layout>
+
+