+++ RazorSvelte/App/AuthorizedPage.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
-    import Layout from "./shared/Layout.svelte";
-    import { getFromJson } from "./shared/config";
+    import Layout from "./Shared/Layout.svelte";
+    import { getFromJson } from "./Shared/Config";
 
     let user = getFromJson<{name: string, email: string, timezone: string, timestamp: string}>("user");
 </script>