+++ RazorSvelte/App/shared/urls.ts
@@ -1,11 +1,11 @@
-﻿import { getFromJson } from "./config";
+﻿import { getFromJson } from "./Config";
 
 const urls = getFromJson<{
     authorizedUrl: string;
     indexUrl: string;
     loginUrl: string;
     logoutUrl: string;
-    privacyUrl: string;
+    aboutUrl: string;
     spaUrl: string;
     signInGoogleUrl: string;
     signInLinkedInUrl: string;