+++ RazorSvelte/Pages/About.entry.ts
@@ -1,4 +1,3 @@
 /// <reference types="svelte" />
-import App from "../App/Privacy.svelte";
-import "bootstrap/js/src/collapse";
+import App from "../App/About.svelte";
 export default new App({target: document.body});