+++ RazorSvelte/Pages/About.rollup.config.js
@@ -1,0 +1,3 @@
+import config from "../Scripts/rollup.config";
+export default config("./Pages/About.entry.ts");
+