+++ RazorSvelte/Pages/Index.entry.ts
@@ -1,10 +1,9 @@
 /// <reference types="svelte" />
 import App from "../App/Index.svelte";
-import "bootstrap/js/src/collapse";
 
 export default new App({
     target: document.body,
     props: {
-        name: "world from svelte"
+        name: ".NET6, ASP.NET, Razor, Svelte, Material UI, TypeScript and SCSS"
     }
 });