+++ RazorSvelte/Pages/Shared/_Loading.cshtml.cs
@@ -4,7 +4,6 @@
     {
         public string Id { get; init; } = "loading";
         public string MarginTop { get; init; } = "100px";
-        public string LoadingClass { get; init; } = "spinner-grow";
         public string Size { get; init; } = "10rem";
     }
 }