+++ RazorSvelte/Scripts/rollup.config.js
@@ -51,7 +51,15 @@ export default (param, globals) => {
                 compilerOptions: {
                     // enable run-time checks when not in production
                     dev: !production
-                }
+                },
+                onwarn: (warning, handler) => {
+                    const { code } = warning;
+                    // skip css unused selector warnings
+                    if (code === "css-unused-selector") {
+                        return;
+                    }
+                    handler(warning);
+                },
             }),
             // we"ll extract any component CSS out into
             // a separate file - better for performance